chore: clean flow types and group the retrieve epics
