+++ src/core_modules/capture-core/components/WorkingLists/TeiWorkingLists/epics/teiViewEpics/index.js
@@ -1,14 +1,9 @@
 // @flow
 export { initTeiViewEpic, updateTeiListEpic } from './teiView.epics';
-export {
-    retrieveTEITemplatesEpic,
-    updateTEITemplateEpic,
-    addTEITemplateEpic,
-    deleteTEITemplateEpic,
-} from './teiTemplates.epics';
+export { updateTEITemplateEpic, addTEITemplateEpic, deleteTEITemplateEpic } from './teiTemplates.epics';
 export {
     addProgramStageTemplateEpic,
     deleteProgramStageTemplateEpic,
     updateProgramStageTemplateEpic,
 } from './programStageTemplates.epics';
-export { retrieveAllTemplatesEpic } from './templatesEpic.epics';
+export { retrieveAllTemplatesEpic, retrieveTEITemplatesEpic } from './retrieveTemplates.epics';
